lakera-tutorials/requirements.txt
requests
typing
numpy
openai
tqdm
datasets
python-dotenv
tiktoken
scipy